+++ packages/traffic-sign-converter/src/data-definitions/DE/data/speed_maxspeed_end.ts
@@ -14,8 +14,7 @@ export const _speed_maxspeed_end: SignType[] = [
       defaultValue: '47',
       format: 'integer',
     },
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -34,8 +33,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -55,8 +53,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -75,8 +72,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -96,8 +92,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -117,8 +112,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -138,8 +132,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -159,8 +152,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -180,8 +172,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -201,8 +192,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -222,8 +212,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -243,8 +232,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -264,8 +252,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -285,8 +272,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',
@@ -306,8 +292,7 @@ export const _speed_maxspeed_end: SignType[] = [
     descriptiveName: 'Ende der zulässigen Höchstgeschwindigkeit',
     description: null,
     kind: 'traffic_sign',
-    tagRecommendations: {},
-    taggingSuggestionsQa: 'explicit_none',
+    tagRecommendations: 'none',
     comments: [],
     catalogue: {
       signCategory: 'speed',